+++ android/autodispose-androidx-lifecycle-test/Module.md
@@ -1,3 +1,3 @@
-# Module autodispose-android-archcomponents-test
+# Module autodispose-androidx-lifecycle-test
 
-Test helpers for `autodispose-android-archcomponents` support.
+Test helpers for `autodispose-androidx-lifecycle` support.
