+++ android/autodispose-androidx-lifecycle-test/gradle.properties
@@ -14,7 +14,7 @@
 # limitations under the License.
 #
 
-POM_NAME=AutoDispose (Android Architecture Components Test Extensions)
-POM_ARTIFACT_ID=autodispose-android-archcomponents-test
+POM_NAME=AutoDispose (androidx-lifecycle test extensions)
+POM_ARTIFACT_ID=autodispose-androidx-lifecycle-test
 POM_PACKAGING=aar
-AUTOMATIC_MODULE_NAME=autodispose2.android.lifecycle.test
+AUTOMATIC_MODULE_NAME=autodispose2.androidx.lifecycle.test
