+++ android/autodispose-androidx-lifecycle/gradle.properties
@@ -14,7 +14,7 @@
 # limitations under the License.
 #
 
-POM_NAME=AutoDispose (Android Architecture Components Extensions)
-POM_ARTIFACT_ID=autodispose-android-archcomponents
+POM_NAME=AutoDispose (androidx-lifecycle extensions)
+POM_ARTIFACT_ID=autodispose-androidx-lifecycle
 POM_PACKAGING=aar
-AUTOMATIC_MODULE_NAME=autodispose2.android.lifecycle
+AUTOMATIC_MODULE_NAME=autodispose2.androidx.lifecycle
