+++ android/autodispose-androidx-lifecycle/src/main/AndroidManifest.xml
@@ -14,4 +14,4 @@
   ~ limitations under the License.
   -->
 
-<manifest package="autodispose2.android.lifecycle"/>
+<manifest package="autodispose2.androidx.lifecycle"/>
