+++ android/autodispose-androidx-lifecycle/src/main/java/autodispose2/androidx/lifecycle/AndroidLifecycleScopeProvider.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package autodispose2.android.lifecycle;
+package autodispose2.androidx.lifecycle;
 
 import androidx.lifecycle.Lifecycle;
 import androidx.lifecycle.LifecycleOwner;
