+++ android/autodispose-androidx-lifecycle/src/main/java/autodispose2/androidx/lifecycle/KotlinExtensions.kt
@@ -15,7 +15,7 @@
  */
 @file:Suppress("NOTHING_TO_INLINE")
 
-package autodispose2.android.lifecycle
+package autodispose2.androidx.lifecycle
 
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.Lifecycle.Event
@@ -41,7 +41,8 @@ import io.reactivex.rxjava3.parallel.ParallelFlowable
  * Extension that returns a [ScopeProvider] for this [LifecycleOwner].
  */
 @CheckReturnValue
-inline fun LifecycleOwner.scope(): ScopeProvider = AndroidLifecycleScopeProvider.from(this)
+inline fun LifecycleOwner.scope(): ScopeProvider = AndroidLifecycleScopeProvider.from(
+    this)
 
 /**
  * Extension that returns a [ScopeProvider] for this [LifecycleOwner].
@@ -51,7 +52,8 @@ inline fun LifecycleOwner.scope(): ScopeProvider = AndroidLifecycleScopeProvider
 @CheckReturnValue
 inline fun LifecycleOwner.scope(
   untilEvent: Lifecycle.Event
-): ScopeProvider = AndroidLifecycleScopeProvider.from(this, untilEvent)
+): ScopeProvider = AndroidLifecycleScopeProvider.from(this,
+    untilEvent)
 
 /**
  * Extension that returns a [ScopeProvider] for this [LifecycleOwner].
@@ -61,13 +63,15 @@ inline fun LifecycleOwner.scope(
 @CheckReturnValue
 inline fun LifecycleOwner.scope(
   boundaryResolver: CorrespondingEventsFunction<Event>
-): ScopeProvider = AndroidLifecycleScopeProvider.from(this, boundaryResolver)
+): ScopeProvider = AndroidLifecycleScopeProvider.from(this,
+    boundaryResolver)
 
 /**
  * Extension that returns a [ScopeProvider] for this [Lifecycle].
  */
 @CheckReturnValue
-inline fun Lifecycle.scope(): ScopeProvider = AndroidLifecycleScopeProvider.from(this)
+inline fun Lifecycle.scope(): ScopeProvider = AndroidLifecycleScopeProvider.from(
+    this)
 
 /**
  * Extension that returns a [ScopeProvider] for this [Lifecycle].
@@ -103,7 +107,9 @@ inline fun <T> Flowable<T>.autoDispose(lifecycleOwner: LifecycleOwner, untilEven
   return if (untilEvent == null) {
     this.to(AutoDispose.autoDisposable(AndroidLifecycleScopeProvider.from(lifecycleOwner)))
   } else {
-    this.to(AutoDispose.autoDisposable(AndroidLifecycleScopeProvider.from(lifecycleOwner, untilEvent)))
+    this.to(AutoDispose.autoDisposable(
+        AndroidLifecycleScopeProvider.from(lifecycleOwner,
+            untilEvent)))
   }
 }
 
@@ -117,9 +123,12 @@ inline fun <T> Flowable<T>.autoDispose(lifecycleOwner: LifecycleOwner, untilEven
 @CheckReturnValue
 inline fun <T> Observable<T>.autoDispose(lifecycleOwner: LifecycleOwner, untilEvent: Event? = null): ObservableSubscribeProxy<T> {
   return if (untilEvent == null) {
-    this.to(AutoDispose.autoDisposable(AndroidLifecycleScopeProvider.from(lifecycleOwner)))
+    this.to(AutoDispose.autoDisposable(
+        AndroidLifecycleScopeProvider.from(lifecycleOwner)))
   } else {
-    this.to(AutoDispose.autoDisposable(AndroidLifecycleScopeProvider.from(lifecycleOwner, untilEvent)))
+    this.to(AutoDispose.autoDisposable(
+        AndroidLifecycleScopeProvider.from(lifecycleOwner,
+            untilEvent)))
   }
 }
 
@@ -133,9 +142,12 @@ inline fun <T> Observable<T>.autoDispose(lifecycleOwner: LifecycleOwner, untilEv
 @CheckReturnValue
 inline fun <T> Single<T>.autoDispose(lifecycleOwner: LifecycleOwner, untilEvent: Event? = null): SingleSubscribeProxy<T> {
   return if (untilEvent == null) {
-    this.to(AutoDispose.autoDisposable(AndroidLifecycleScopeProvider.from(lifecycleOwner)))
+    this.to(AutoDispose.autoDisposable(
+        AndroidLifecycleScopeProvider.from(lifecycleOwner)))
   } else {
-    this.to(AutoDispose.autoDisposable(AndroidLifecycleScopeProvider.from(lifecycleOwner, untilEvent)))
+    this.to(AutoDispose.autoDisposable(
+        AndroidLifecycleScopeProvider.from(lifecycleOwner,
+            untilEvent)))
   }
 }
 
@@ -149,9 +161,12 @@ inline fun <T> Single<T>.autoDispose(lifecycleOwner: LifecycleOwner, untilEvent:
 @CheckReturnValue
 inline fun <T> Maybe<T>.autoDispose(lifecycleOwner: LifecycleOwner, untilEvent: Event? = null): MaybeSubscribeProxy<T> {
   return if (untilEvent == null) {
-    this.to(AutoDispose.autoDisposable(AndroidLifecycleScopeProvider.from(lifecycleOwner)))
+    this.to(AutoDispose.autoDisposable(
+        AndroidLifecycleScopeProvider.from(lifecycleOwner)))
   } else {
-    this.to(AutoDispose.autoDisposable(AndroidLifecycleScopeProvider.from(lifecycleOwner, untilEvent)))
+    this.to(AutoDispose.autoDisposable(
+        AndroidLifecycleScopeProvider.from(lifecycleOwner,
+            untilEvent)))
   }
 }
 
@@ -165,9 +180,12 @@ inline fun <T> Maybe<T>.autoDispose(lifecycleOwner: LifecycleOwner, untilEvent: 
 @CheckReturnValue
 inline fun Completable.autoDispose(lifecycleOwner: LifecycleOwner, untilEvent: Event? = null): CompletableSubscribeProxy {
   return if (untilEvent == null) {
-    this.to(AutoDispose.autoDisposable<Any>(AndroidLifecycleScopeProvider.from(lifecycleOwner)))
+    this.to(AutoDispose.autoDisposable<Any>(
+        AndroidLifecycleScopeProvider.from(lifecycleOwner)))
   } else {
-    this.to(AutoDispose.autoDisposable<Any>(AndroidLifecycleScopeProvider.from(lifecycleOwner, untilEvent)))
+    this.to(AutoDispose.autoDisposable<Any>(
+        AndroidLifecycleScopeProvider.from(lifecycleOwner,
+            untilEvent)))
   }
 }
 
@@ -181,8 +199,11 @@ inline fun Completable.autoDispose(lifecycleOwner: LifecycleOwner, untilEvent: E
 @CheckReturnValue
 inline fun <T> ParallelFlowable<T>.autoDispose(lifecycleOwner: LifecycleOwner, untilEvent: Event? = null): ParallelFlowableSubscribeProxy<T> {
   return if (untilEvent == null) {
-    this.to(AutoDispose.autoDisposable(AndroidLifecycleScopeProvider.from(lifecycleOwner)))
+    this.to(AutoDispose.autoDisposable(
+        AndroidLifecycleScopeProvider.from(lifecycleOwner)))
   } else {
-    this.to(AutoDispose.autoDisposable(AndroidLifecycleScopeProvider.from(lifecycleOwner, untilEvent)))
+    this.to(AutoDispose.autoDisposable(
+        AndroidLifecycleScopeProvider.from(lifecycleOwner,
+            untilEvent)))
   }
 }
