+++ sample/src/main/kotlin/autodispose2/sample/ArchComponentActivity.kt
@@ -22,7 +22,7 @@ import android.widget.ImageView
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.ViewModelProviders
 import autodispose2.autoDispose
-import autodispose2.android.lifecycle.AndroidLifecycleScopeProvider
+import autodispose2.androidx.lifecycle.AndroidLifecycleScopeProvider
 import autodispose2.sample.repository.ImageRepository
 import io.reactivex.rxjava3.android.schedulers.AndroidSchedulers
 import io.reactivex.rxjava3.schedulers.Schedulers
