+++ sample/src/main/kotlin/autodispose2/sample/DisposingViewModelActivity.kt
@@ -22,7 +22,7 @@ import android.widget.ProgressBar
 import android.widget.TextView
 import autodispose2.ScopeProvider
 import autodispose2.autoDispose
-import autodispose2.android.lifecycle.AndroidLifecycleScopeProvider
+import autodispose2.androidx.lifecycle.AndroidLifecycleScopeProvider
 import autodispose2.sample.repository.NetworkRepository
 import autodispose2.sample.state.DownloadState
 import io.reactivex.rxjava3.android.schedulers.AndroidSchedulers
